+++ Basil/include/Renderer/Renderer2D.h
@@ -36,6 +36,8 @@ namespace Basil
 			static void drawSprite(const glm::mat4& transform, SpriteRendererComponent& src, int entityID);
 			static void drawCircle(const glm::mat4& transform, const glm::vec4& color, float thickness = 1.0f, float fade = 0.005f, int entityID = -1);
 			static void drawLine(const glm::vec3& p0, const glm::vec3& p1, const glm::vec4& color, int entityID = -1);
+			static void drawRect(const glm::vec3& position, const glm::vec2& size, const glm::vec4& color, int entityID = -1);
+			static void drawRect(const glm::mat4& transform, const glm::vec4& color, int entityID = -1);
 			static float getLineWidth();
 			static void setLineWidth(float width);